Move flac_write_picture to flac_picture.c
This change moves flac_write_picture to flac_picture.c To match the naming convention, the function was renamed to ff_flac_write_picture.

The function was modified to only write the METADATA_BLOCK_PICTURE structure. The code to write the METADATA_BLOCK_HEADER structure was moved into the function flac_finish_header in flacenc.c. To facilitate this move, ff_flac_picture_length was added to flac_picture.c, which returns the length of the number of bytes that ff_flac_write_picture will write. An argument of type AVIOContext* was added to ff_flac_write_picture to allow for writing to other locations, such as to a memory stream for base64 encoding.

Signed-off-by: Lewis Fox <LRFLEW@aol.com>